requirements.txt
langchain-openai
langchain-community
langchain-core
python-dotenv
faiss-cpu